finished unit testing

+++ migrations/sqls/20220910142720-orders-table-up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE orders(
     id SERIAL PRIMARY KEY, 
-    user_id BIGINT REFERENCES users(id), 
+    user_id INT REFERENCES users(id), 
     is_complete BOOLEAN NOT NULL
 );
